+++ tests/unit_tests/test_tethys_apps/test_cli/test_services_commands.py
@@ -5,8 +5,10 @@ except ImportError:
 import unittest
 from unittest import mock
 
-from tethys_apps.cli.services_commands import services_create_persistent_command, services_remove_persistent_command,\
-    services_create_spatial_command, services_remove_spatial_command, services_list_command
+from tethys_apps.cli.services_commands import (services_create_persistent_command, services_remove_persistent_command,
+                                               services_create_spatial_command, services_remove_spatial_command,
+                                               services_list_command, services_create_dataset_command,
+                                               services_remove_dataset_command, services_create_wps_command)
 from django.core.exceptions import ObjectDoesNotExist
 from django.db.utils import IntegrityError
 
@@ -38,7 +40,6 @@ class ServicesCommandsTest(unittest.TestCase):
         """
         mock_args = mock.MagicMock()
         services_create_persistent_command(mock_args)
-
         mock_service.assert_called()
 
         po_call_args = mock_pretty_output().__enter__().write.call_args_list
@@ -100,7 +101,9 @@ class ServicesCommandsTest(unittest.TestCase):
         :param mock_pretty_output: mock for pretty_output text
         :return:
         """
+
         mock_args = mock.MagicMock()
+        mock_service.__str__.return_value = 'Persistent Store'
         mock_args.force = True
         mock_service.objects.get.side_effect = [ValueError, ObjectDoesNotExist]
         # NOTE: to prevent our tests from exiting prematurely, we change the behavior of exit to raise an exception
@@ -128,6 +131,8 @@ class ServicesCommandsTest(unittest.TestCase):
         """
         mock_args = mock.MagicMock()
         mock_args.force = True
+        mock_service.__str__.return_value = 'Persistent Store'
+
         # NOTE: to prevent our tests from exiting prematurely, we change the behavior of exit to raise an exception
         # to break the code execution, which we catch below.
         mock_exit.side_effect = SystemExit
@@ -161,6 +166,7 @@ class ServicesCommandsTest(unittest.TestCase):
         # to break the code execution, which we catch below.
         mock_exit.side_effect = SystemExit
         mock_input.side_effect = ['foo', 'N']
+        mock_service.__str__.return_value = 'Persistent Store'
 
         self.assertRaises(SystemExit, services_remove_persistent_command, mock_args)
 
@@ -172,8 +178,8 @@ class ServicesCommandsTest(unittest.TestCase):
 
         po_call_args = mock_input.call_args_list
         self.assertEqual(2, len(po_call_args))
-        self.assertEqual('Are you sure you want to delete this Persistent Store Service? [y/n]: ',
-                         po_call_args[0][0][0])
+        self.assertEqual(
+            'Are you sure you want to delete this Persistent Store Service? [y/n]: ', po_call_args[0][0][0])
         self.assertEqual('Please enter either "y" or "n": ', po_call_args[1][0][0])
 
     @mock.patch('tethys_apps.cli.services_commands.input')
@@ -191,6 +197,7 @@ class ServicesCommandsTest(unittest.TestCase):
         :return:
         """
         mock_args = mock.MagicMock()
+        mock_service.__str__.return_value = 'Persistent Store'
         mock_args.force = False
         # NOTE: to prevent our tests from exiting prematurely, we change the behavior of exit to raise an exception
         # to break the code execution, which we catch below.
@@ -207,8 +214,8 @@ class ServicesCommandsTest(unittest.TestCase):
 
         po_call_args = mock_input.call_args_list
         self.assertEqual(1, len(po_call_args))
-        self.assertEqual('Are you sure you want to delete this Persistent Store Service? [y/n]: ',
-                         po_call_args[0][0][0])
+        self.assertEqual(
+            'Are you sure you want to delete this Persistent Store Service? [y/n]: ', po_call_args[0][0][0])
 
     @mock.patch('tethys_apps.cli.services_commands.pretty_output')
     @mock.patch('tethys_services.models.SpatialDatasetService')
@@ -315,6 +322,8 @@ class ServicesCommandsTest(unittest.TestCase):
         :return:
         """
         mock_args = mock.MagicMock()
+        mock_service.__str__.return_value = 'Spatial Dataset'
+
         mock_service.objects.get.side_effect = [ValueError, ObjectDoesNotExist]
         # NOTE: to prevent our tests from exiting prematurely, we change the behavior of exit to raise an exception
         # to break the code execution, which we catch below.
@@ -340,6 +349,8 @@ class ServicesCommandsTest(unittest.TestCase):
         :return:
         """
         mock_args = mock.MagicMock()
+        mock_service.__str__.return_value = 'Spatial Dataset'
+
         mock_args.force = True
         # NOTE: to prevent our tests from exiting prematurely, we change the behavior of exit to raise an exception
         # to break the code execution, which we catch below.
@@ -369,6 +380,8 @@ class ServicesCommandsTest(unittest.TestCase):
         :return:
         """
         mock_args = mock.MagicMock()
+        mock_service.__str__.return_value = 'Spatial Dataset'
+
         mock_args.force = False
         # NOTE: to prevent our tests from exiting prematurely, we change the behavior of exit to raise an exception
         # to break the code execution, which we catch below.
@@ -385,8 +398,7 @@ class ServicesCommandsTest(unittest.TestCase):
 
         po_call_args = mock_input.call_args_list
         self.assertEqual(2, len(po_call_args))
-        self.assertEqual('Are you sure you want to delete this Persistent Store Service? [y/n]: ',
-                         po_call_args[0][0][0])
+        self.assertEqual('Are you sure you want to delete this Spatial Dataset Service? [y/n]: ', po_call_args[0][0][0])
         self.assertEqual('Please enter either "y" or "n": ', po_call_args[1][0][0])
 
     @mock.patch('tethys_apps.cli.services_commands.input')
@@ -404,6 +416,8 @@ class ServicesCommandsTest(unittest.TestCase):
         :return:
         """
         mock_args = mock.MagicMock()
+        mock_service.__str__.return_value = 'Spatial Dataset'
+
         mock_args.force = False
         # NOTE: to prevent our tests from exiting prematurely, we change the behavior of exit to raise an exception
         # to break the code execution, which we catch below.
@@ -420,8 +434,7 @@ class ServicesCommandsTest(unittest.TestCase):
 
         po_call_args = mock_input.call_args_list
         self.assertEqual(1, len(po_call_args))
-        self.assertEqual('Are you sure you want to delete this Persistent Store Service? [y/n]: ',
-                         po_call_args[0][0][0])
+        self.assertEqual('Are you sure you want to delete this Spatial Dataset Service? [y/n]: ', po_call_args[0][0][0])
 
     @mock.patch('tethys_apps.cli.services_commands.print')
     @mock.patch('tethys_apps.cli.services_commands.pretty_output')
@@ -444,6 +457,8 @@ class ServicesCommandsTest(unittest.TestCase):
         mock_args = mock.MagicMock()
         mock_args.spatial = False
         mock_args.persistent = False
+        mock_args.dataset = False
+        mock_args.wps = False
         mock_spatial.objects.order_by('id').all.return_value = [mock.MagicMock(), mock.MagicMock(), mock.MagicMock()]
         mock_persistent.objects.order_by('id').all.return_value = [mock.MagicMock(), mock.MagicMock()]
 
@@ -501,6 +516,8 @@ class ServicesCommandsTest(unittest.TestCase):
         mock_args = mock.MagicMock()
         mock_args.spatial = True
         mock_args.persistent = False
+        mock_args.dataset = False
+        mock_args.wps = False
         mock_spatial.objects.order_by('id').all.return_value = [mock.MagicMock(), mock.MagicMock(), mock.MagicMock()]
 
         services_list_command(mock_args)
@@ -546,6 +563,8 @@ class ServicesCommandsTest(unittest.TestCase):
         mock_args = mock.MagicMock()
         mock_args.spatial = False
         mock_args.persistent = True
+        mock_args.dataset = False
+        mock_args.wps = False
         mock_persistent.objects.order_by('id').all.return_value = [mock.MagicMock(), mock.MagicMock()]
 
         services_list_command(mock_args)
@@ -572,3 +591,144 @@ class ServicesCommandsTest(unittest.TestCase):
         self.assertNotIn(self.my_dict['endpoint'], rts_call_args[1][0][0])
         self.assertNotIn(self.my_dict['public_endpoint'], rts_call_args[1][0][0])
         self.assertNotIn(self.my_dict['apikey'], rts_call_args[1][0][0])
+
+    @mock.patch('tethys_apps.cli.services_commands.pretty_output')
+    @mock.patch('tethys_services.models.DatasetService')
+    def test_services_create_dataset_command_IndexError(self, mock_service, mock_pretty_output):
+
+        mock_args = mock.MagicMock()
+        mock_args.connection = 'IndexError:9876@IndexError'  # No 'http' or '://'
+
+        services_create_dataset_command(mock_args)
+
+        mock_service.assert_not_called()
+
+        po_call_args = mock_pretty_output().__enter__().write.call_args_list
+        self.assertEqual(1, len(po_call_args))
+        self.assertIn('The connection argument (-c) must be of the form', po_call_args[0][0][0])
+        self.assertIn('"<username>:<password>@<protocol>//<host>:<port>".', po_call_args[0][0][0])
+
+    @mock.patch('tethys_apps.cli.services_commands.pretty_output')
+    @mock.patch('tethys_services.models.DatasetService')
+    def test_services_create_dataset_command_FormatError(self, mock_service, mock_pretty_output):
+
+        mock_args = mock.MagicMock()
+        mock_args.connection = 'foo:pass@http:://foo:1234'
+        mock_args.public_endpoint = 'foo@foo:foo'  # No 'http' or '://'
+
+        services_create_dataset_command(mock_args)
+
+        mock_service.assert_not_called()
+
+        po_call_args = mock_pretty_output().__enter__().write.call_args_list
+        self.assertEqual(1, len(po_call_args))
+        self.assertIn('The public_endpoint argument (-p) must be of the form ', po_call_args[0][0][0])
+        self.assertIn('"<protocol>//<host>:<port>".', po_call_args[0][0][0])
+
+    @mock.patch('tethys_apps.cli.services_commands.pretty_output')
+    @mock.patch('tethys_services.models.DatasetService')
+    def test_services_create_dataset_command_IntegrityError(self, mock_service, mock_pretty_output):
+
+        mock_args = mock.MagicMock()
+        mock_args.connection = 'foo:pass@http:://foo:1234'
+        mock_args.public_endpoint = 'http://foo:1234'
+        mock_service.side_effect = IntegrityError
+
+        services_create_dataset_command(mock_args)
+
+        mock_service.assert_called()
+
+        po_call_args = mock_pretty_output().__enter__().write.call_args_list
+        self.assertEqual(1, len(po_call_args))
+        self.assertIn('Dataset Service with name ', po_call_args[0][0][0])
+        self.assertIn('already exists. Command aborted.', po_call_args[0][0][0])
+
+    @mock.patch('tethys_apps.cli.services_commands.pretty_output')
+    @mock.patch('tethys_services.models.DatasetService')
+    def test_services_create_dataset_command(self, mock_service, mock_pretty_output):
+        mock_args = mock.MagicMock()
+        mock_args.connection = 'foo:pass@http:://foo:1234'
+        mock_args.public_endpoint = 'http://foo:1234'
+        mock_service.return_value = mock.MagicMock()
+
+        services_create_dataset_command(mock_args)
+
+        mock_service.assert_called()
+
+        po_call_args = mock_pretty_output().__enter__().write.call_args_list
+        self.assertEqual(1, len(po_call_args))
+        self.assertEqual('Successfully created new Dataset Service!', po_call_args[0][0][0])
+
+    @mock.patch('tethys_apps.cli.services_commands.input')
+    @mock.patch('tethys_apps.cli.services_commands.pretty_output')
+    @mock.patch('tethys_apps.cli.services_commands.exit')
+    @mock.patch('tethys_services.models.DatasetService')
+    def test_services_remove_dataset_command_proceed(self, mock_service, mock_exit, mock_pretty_output, mock_input):
+
+        mock_args = mock.MagicMock()
+        mock_service.__str__.return_value = 'Dataset'
+
+        mock_args.force = False
+        mock_exit.side_effect = SystemExit
+        mock_input.side_effect = ['y']
+
+        self.assertRaises(SystemExit, services_remove_dataset_command, mock_args)
+
+        mock_service.objects.get().delete.assert_called()
+
+        po_call_args = mock_pretty_output().__enter__().write.call_args_list
+        self.assertEqual(1, len(po_call_args))
+        self.assertIn('Successfully removed Dataset Service', po_call_args[0][0][0])
+
+        po_call_args = mock_input.call_args_list
+        self.assertEqual(1, len(po_call_args))
+        self.assertEqual('Are you sure you want to delete this Dataset Service? [y/n]: ', po_call_args[0][0][0])
+
+    @mock.patch('tethys_apps.cli.services_commands.pretty_output')
+    @mock.patch('tethys_services.models.WebProcessingService')
+    def test_services_create_wps_command_IndexError(self, mock_service, mock_pretty_output):
+
+        mock_args = mock.MagicMock()
+        mock_args.connection = 'IndexError:9876@IndexError'  # No 'http' or '://'
+
+        services_create_wps_command(mock_args)
+
+        mock_service.assert_not_called()
+
+        po_call_args = mock_pretty_output().__enter__().write.call_args_list
+        self.assertEqual(1, len(po_call_args))
+        self.assertIn('The connection argument (-c) must be of the form', po_call_args[0][0][0])
+        self.assertIn('"<username>:<password>@<protocol>//<host>:<port>".', po_call_args[0][0][0])
+
+    @mock.patch('tethys_apps.cli.services_commands.pretty_output')
+    @mock.patch('tethys_services.models.WebProcessingService')
+    def test_services_create_wps_command_IntegrityError(self, mock_service, mock_pretty_output):
+
+        mock_args = mock.MagicMock()
+        mock_args.connection = 'foo:pass@http:://foo:1234'
+        mock_args.public_endpoint = 'http://foo:1234'
+        mock_service.side_effect = IntegrityError
+
+        services_create_wps_command(mock_args)
+
+        mock_service.assert_called()
+
+        po_call_args = mock_pretty_output().__enter__().write.call_args_list
+        self.assertEqual(1, len(po_call_args))
+        self.assertIn('Web Processing Service with name ', po_call_args[0][0][0])
+        self.assertIn('already exists. Command aborted.', po_call_args[0][0][0])
+
+    @mock.patch('tethys_apps.cli.services_commands.pretty_output')
+    @mock.patch('tethys_services.models.WebProcessingService')
+    def test_services_create_wps_command(self, mock_service, mock_pretty_output):
+        mock_args = mock.MagicMock()
+        mock_args.connection = 'foo:pass@http:://foo:1234'
+        mock_service.return_value = mock.MagicMock()
+
+        services_create_wps_command(mock_args)
+
+        mock_service.assert_called()
+
+        po_call_args = mock_pretty_output().__enter__().write.call_args_list
+        self.assertEqual(1, len(po_call_args))
+        self.assertEqual('Successfully created new Web Processing Service!', po_call_args[0][0][0])