+++ tests/unit_tests/test_tethys_apps/test_harvester.py
@@ -154,7 +154,7 @@ class HarvesterTest(unittest.TestCase):
         :return:
         """
         mock_args = mock.MagicMock()
-        list_apps = ['foo']
+        list_apps = {'foo': 'foo'}
         mock_args = list_apps
 
         shv = SingletonHarvester()
@@ -178,7 +178,8 @@ class HarvesterTest(unittest.TestCase):
         :return:
         """
         mock_args = mock.MagicMock()
-        list_apps = [u'.gitignore', u'test_app', u'__init__.py', u'__init__.pyc']
+        list_apps = {'test_app': 'tethysapp.test_app'}
+
         mock_args = list_apps
         mock_subclass.side_effect = TypeError
 
@@ -194,7 +195,7 @@ class HarvesterTest(unittest.TestCase):
 
     @mock.patch('sys.stdout', new_callable=io.StringIO)
     @mock.patch('tethys_apps.harvester.tethys_log.exception')
-    @mock.patch('tethys_apps.tethysapp.test_app.app.TestApp.url_maps')
+    @mock.patch('tethysapp.test_app.app.TestApp.url_maps')
     def test_harvest_app_instances_Exceptions1(self, mock_url_maps, mock_logexception, mock_stdout):
         """
         Test for SingletonHarvester._harvest_app_instances
@@ -206,7 +207,7 @@ class HarvesterTest(unittest.TestCase):
         :return:
         """
         mock_args = mock.MagicMock()
-        list_apps = [u'.gitignore', u'test_app', u'__init__.py', u'__init__.pyc']
+        list_apps = {'test_app': 'tethysapp.test_app'}
         mock_args = list_apps
         mock_url_maps.side_effect = ImportError
 
@@ -219,7 +220,7 @@ class HarvesterTest(unittest.TestCase):
 
     @mock.patch('sys.stdout', new_callable=io.StringIO)
     @mock.patch('tethys_apps.harvester.tethys_log.warning')
-    @mock.patch('tethys_apps.tethysapp.test_app.app.TestApp.register_app_permissions')
+    @mock.patch('tethysapp.test_app.app.TestApp.register_app_permissions')
     def test_harvest_app_instances_Exceptions2(self, mock_permissions, mock_logwarning, mock_stdout):
         """
         Test for SingletonHarvester._harvest_app_instances
@@ -230,7 +231,8 @@ class HarvesterTest(unittest.TestCase):
         :param mock_stdout:  mock for the text output
         :return:
         """
-        list_apps = [u'.gitignore', u'test_app', u'__init__.py', u'__init__.pyc']
+        list_apps = {'test_app': 'tethysapp.test_app'}
+
         mock_permissions.side_effect = ProgrammingError
 
         shv = SingletonHarvester()