+++ tests/unit_tests/test_tethys_apps/test_helpers.py
@@ -22,17 +22,6 @@ class TethysAppsHelpersTests(unittest.TestCase):
         result = helpers.get_installed_tethys_apps()
         self.assertTrue('test_app' in result)
 
-    @mock.patch('tethys_apps.helpers.os.path.isdir')
-    @mock.patch('tethys_apps.helpers.os.listdir')
-    @mock.patch('tethys_apps.helpers.get_tethysapp_dir')
-    def test_get_installed_tethys_apps_mock(self, mock_dir, mock_listdir, mock_isdir):
-        # Get list of apps installed in the mock directory
-        mock_dir.return_value = '/foo/bar'
-        mock_listdir.return_value = ['.gitignore', 'foo_app', '__init__.py', '__init__.pyc']
-        mock_isdir.side_effect = [False, True, False, False]
-        result = helpers.get_installed_tethys_apps()
-        self.assertTrue('foo_app' in result)
-
     def test_get_installed_tethys_extensions(self):
         # Get a list of installed extensions
         result = helpers.get_installed_tethys_extensions()